src/AI/viewer/requirements.txt
Pillow
natsort==7.1.0
